Use standard-library that's already provided by nix
Instead of copying from the nix store point Agda where to look for
standard library.  Turning on --local-interfaces is necessary because
otherwise agda will try to write stuff in the nix store.

Processed standard library sources are in the `_build/html0`, so for
now I just copy them.  Navigating to them will display page that is
not so pretty, but I think it's better than 404.

To do something nicer with them, we would probably have to write some
shake rules and instruct how to process them further.